src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/shadow/svg-style-in-shadow-tree-expected.txt

CONSOLE ERROR: line 19: Uncaught TypeError: document.registerElement is not a function
